Refactor: Move InputEventFrames out of GameState
Moves `InputEventFrames` list and logic from `GameState` to `MainContext` (and other game contexts like `ServerContext` and `TerrainSandboxContext`). Updates `Structure.maybeInteract()` to take an event consumer rather than `GameState`, and updates tests accordingly.
