+++ nomad-realms-server/src/main/java/nomadrealms/app/context/ServerContext.java
@@ -10,6 +10,7 @@ import engine.context.input.event.PacketReceivedInputEvent;
 import engine.networking.NetworkingReceiver;
 import nomadrealms.context.game.GameState;
 import nomadrealms.context.game.event.InputEvent;
+import nomadrealms.context.game.event.InputEventFrame;
 import nomadrealms.context.game.world.map.generation.OverworldGenerationStrategy;
 import nomadrealms.event.networking.SyncedEvent;
 import nomadrealms.render.RenderingEnvironment;
@@ -19,6 +20,7 @@ public class ServerContext extends GameContext {
 	private RenderingEnvironment re;
 	private GameState gameState;
 	private final Queue<InputEvent> uiEventChannel = new ArrayDeque<>();
+	private final java.util.List<InputEventFrame> inputFrames = new java.util.ArrayList<>();
 
 	private final NetworkingReceiver networkingReceiver = new NetworkingReceiver();
 
@@ -34,7 +36,11 @@ public class ServerContext extends GameContext {
 		if (gameState == null) {
 			return;
 		}
-		gameState.update();
+		inputFrames.add(new InputEventFrame(gameState.frameNumber + 1));
+		if (inputFrames.size() > 30) {
+			inputFrames.remove(0);
+		}
+		gameState.update(inputFrames.get(inputFrames.size() - 1));
 		while (!uiEventChannel.isEmpty()) {
 			uiEventChannel.poll();
 		}
