+++ nomad-realms/src/test/java/nomadrealms/context/game/TerrainSandboxTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 
 import java.util.LinkedList;
 import nomadrealms.context.game.world.map.generation.OverworldGenerationStrategy;
+import nomadrealms.context.game.event.InputEventFrame;
 import nomadrealms.context.game.world.map.generation.TerrainSandboxMapInitialization;
 import org.junit.jupiter.api.Test;
 
@@ -17,7 +18,7 @@ public class TerrainSandboxTest {
 
 		assertNull(gameState.world.nomad);
 		assertDoesNotThrow(() -> gameState.reindex(new LinkedList<>()));
-		assertDoesNotThrow(gameState::update);
+		assertDoesNotThrow(() -> gameState.update(new InputEventFrame(gameState.frameNumber + 1)));
 	}
 
 }
